requirements.txt
streamlit
pillow
google-cloud
google-cloud-vision